requirements.txt
fastapi
uvicorn

langchain>=0.1.0
langchain-community>=0.0.30

transformers
sentence-transformers
faiss-cpu
torch

unstructured
pdfminer.six
pypandoc
python-docx
python-pptx
openpyxl
pillow
